+++ application/controllers/Edit.php
@@ -2,27 +2,24 @@
 
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-class MaintainSet extends Application {
+class Edit extends Application {
 
     //put your code here
-    public function Add() {
+    public function Initialize($id) {
         $role = $this->session->userdata('userrole');
         if ($role == ROLE_GUEST) {
             redirect($_SERVER['HTTP_REFERER']); // back where we came from
             return;
         }
-
         $items = $this->Accessories->all();
-
-        $h = array();
-        $c = array();
-        $p = array();
-        $s = array();
-
+        $set = $this->Sets->get($id);
         for ($x = 0; $x < sizeof($items); $x++) {
             $acc = $this->Accessories->get($x);
-            $temp = array('item' => '<option value="' . $acc->id . '">' . $acc->name . '</option>');
-
+            if ($set->helmId == $acc->id || $set->chestId == $acc->id || $set->primaryId == $acc->id || $set->secondaryId == $acc->id) {
+                $temp = array('item' => '<option selected value="' . $acc->id . '">' . $acc->name . '</option>');
+            } else {
+                $temp = array('item' => '<option value="' . $acc->id . '">' . $acc->name . '</option>');
+            }
             if ($acc->catId == 0) {
                 $h[] = $temp;
             }
@@ -36,21 +33,18 @@ class MaintainSet extends Application {
                 $s[] = $temp;
             }
         }
-
         $this->data['h'] = $h;
         $this->data['c'] = $c;
         $this->data['p'] = $p;
         $this->data['s'] = $s;
-
-
-        $this->data['setNum'] = sizeof($items = $this->Sets->all()) + 1;
-        $this->data['pagetitle'] = 'Add Set';
-        $this->data['pagebody'] = 'addSet';
-
+        $this->data['setId'] = $id;
+        $this->data['setName'] = $set->name;
+        $this->data['pagetitle'] = 'Edit Set';
+        $this->data['pagebody'] = 'editSet';
         $this->render();
     }
 
-    public function AddSet() {
+    public function OnSubmit() {
         $role = $this->session->userdata('userrole');
         if ($role == ROLE_GUEST) {
             redirect($_SERVER['HTTP_REFERER']); // back where we came from
@@ -59,17 +53,14 @@ class MaintainSet extends Application {
         $this->load->model('Sets');
         $this->load->library('form_validation');
         $this->form_validation->set_rules($this->Sets->rules());
+        $set = $this->input->post();
+        $set = (object) $set;
 
-        $newSet = $this->input->post();
-        var_dump($newSet);
-        $newSet = (object) $newSet;
-        
         if ($this->form_validation->run()) {
-            $this->Sets->add($newSet);
+            $this->Sets->update($set);
         } else {
             $this->alert(validation_errors(), 'danger');
         }
-        redirect('/Welcome/set/' . $this->Sets->highest());
+        redirect('/Welcome/set/' . $set->id);
     }
-
 }