+++ application/views/editSet.php
@@ -1,15 +1,14 @@
 <div class="container">
     <div class="row">
         <div>
-            <h1>Add a set</h1>
+            <h1>Edit a set</h1>
         </div>
     </div>
-
     <div>
-        <form action="OnSubmit" method="post">
+        <form action="/Edit/OnSubmit" method="post">
             <input type="hidden" name="id" value="{setId}" />
                 <label for="addName">Name</label>
-                <input type="text" name="n"><br>
+                <input type="text" name="n" value="{setName}"><br>
                 <label for="addHelm">Helm</label>
                 <select name="h" id="addHelm" class="form-control">
                     <option selected></option>
@@ -39,8 +38,8 @@
                     {/s}
                 </select>
             </div>
-            <button type="submit" class="btn btn-success">Add</button>
+            <button type="submit" class="btn btn-success">Edit</button>
         </form>
-        <button onclick="location.href = '/welcome';" id="myButton" class="btn btn-danger" >Cancel</button>
+        <button onclick="location.href = '/welcome/';" id="myButton" class="btn btn-danger" >Cancel</button>
     </div>
 </div>